+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## staged
 
+- Added SOC transformer relaxations
 - Fixed bugs in center-tap transformer modeling
 - Add wye-connected CapControl for IVR and FOT (polar) formulations
 - Fixed indexing issue for single-phase delta load models in linear formulations (LinDist3Flow, FOTP, FOTR, FBS)